requirements.txt
streamlit>=1.37
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
Pillow>=10.0
pdfplumber>=0.11
python-docx>=1.1
fastapi>=0.115
uvicorn>=0.30
